backend/requirements.txt
Flask==2.3.3
gunicorn==21.2.0
psycopg2-binary==2.9.9
prometheus-flask-exporter==0.23.0
